requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
jinja2==3.0.1
aiofiles==0.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
sqlalchemy==1.4.23
pandas==1.3.3
openpyxl==3.0.9
pillow==8.3.2
pytesseract==0.3.8
opencv-python==4.5.3.56
numpy==1.21.2 